Add ft_cat with a way to manually test it
