+++ define/define_macosx.s
@@ -6,11 +6,12 @@
 ;    By: jlagneau <jlagneau@student.42.fr>          +#+  +:+       +#+         ;
 ;                                                 +#+#+#+#+#+   +#+            ;
 ;    Created: 2015/10/31 19:38:19 by jlagneau          #+#    #+#              ;
-;    Updated: 2017/03/29 20:38:52 by jlagneau         ###   ########.fr        ;
+;    Updated: 2017/04/02 19:44:05 by jlagneau         ###   ########.fr        ;
 ;                                                                              ;
 ;******************************************************************************;
 
 %include "define/define_global.s"
 
 ;; MAC OS
+%define SYS_READ		0x2000003
 %define SYS_WRITE       0x2000004
