+++ define/macro.s
@@ -6,7 +6,7 @@
 ;    By: jlagneau <jlagneau@student.42.fr>          +#+  +:+       +#+         ;
 ;                                                 +#+#+#+#+#+   +#+            ;
 ;    Created: 2017/03/21 18:44:12 by jlagneau          #+#    #+#              ;
-;    Updated: 2017/03/29 19:52:10 by jlagneau         ###   ########.fr        ;
+;    Updated: 2017/04/02 19:51:16 by jlagneau         ###   ########.fr        ;
 ;                                                                              ;
 ;******************************************************************************;
 
@@ -14,6 +14,15 @@
 %macro sys_write 3
     mov   rax, SYS_WRITE
     mov   rdi, %1
+    mov   rsi, %2
+    mov   rdx, %3
+    syscall
+%endmacro
+
+; Implements the read system call
+%macro sys_read 3
+    mov	  rax, SYS_READ
+    mov	  rdi, %1
     mov   rsi, %2
     mov   rdx, %3
     syscall
